auth/requirements.txt
python-jose[cryptography]==3.3.0
pymongo>=4.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
pydantic>=2.0.0
